+++ R/kernelshap.R
@@ -3,22 +3,16 @@
 #' @description
 #' Efficient implementation of Kernel SHAP, see Lundberg and Lee (2017), and
 #' Covert and Lee (2021), abbreviated by CL21.
-#' By default, for up to p=8 features, exact Kernel SHAP values are returned
+#' By default, for up to p=8 features, exact SHAP values are returned
 #' (with respect to the selected background data).
-#' Otherwise, an almost exact hybrid algorithm combining exact calculations and
+#' Otherwise, a partly exact hybrid algorithm combining exact calculations and
 #' iterative paired sampling is used, see Details.
-#'
-#' Note that (exact) Kernel SHAP is only an approximation of (exact) permutation SHAP.
-#' Thus, for up to eight features, we recommend [permshap()]. For more features,
-#' [permshap()] tends to be inefficient compared the optimized hybrid strategy
-#' of Kernel SHAP.
 #'
 #' @details
 #' The pure iterative Kernel SHAP sampling as in Covert and Lee (2021) works like this:
 #'
-#' 1. A binary "on-off" vector \eqn{z} is drawn from \eqn{\{0, 1\}^p}
-#'   such that its sum follows the SHAP Kernel weight distribution
-#'   (normalized to the range \eqn{\{1, \dots, p-1\}}).
+#' 1. A binary "on-off" vector \eqn{z} is drawn from \eqn{\{0, 1\}^p} according to
+#'   a special weighting logic.
 #' 2. For each \eqn{j} with \eqn{z_j = 1}, the \eqn{j}-th column of the
 #'   original background data is replaced by the corresponding feature value \eqn{x_j}
 #'   of the observation to be explained.
@@ -33,17 +27,13 @@
 #'
 #' This is repeated multiple times until convergence, see CL21 for details.
 #'
-#' A drawback of this strategy is that many (at least 75%) of the \eqn{z} vectors will
-#' have \eqn{\sum z \in \{1, p-1\}}, producing many duplicates. Similarly, at least 92%
-#' of the mass will be used for the \eqn{p(p+1)} possible vectors with
-#' \eqn{\sum z \in \{1, 2, p-2, p-1\}}.
-#' This inefficiency can be fixed by a hybrid strategy, combining exact calculations
-#' with sampling.
+#' To avoid the re-evaluation of identical coalition vectors, we have implemented
+#' a hybrid strategy, combining exact calculations with sampling.
 #'
 #' The hybrid algorithm has two steps:
 #' 1. Step 1 (exact part): There are \eqn{2p} different on-off vectors \eqn{z} with
-#'   \eqn{\sum z \in \{1, p-1\}}, covering a large proportion of the Kernel SHAP
-#'   distribution. The degree 1 hybrid will list those vectors and use them according
+#'   \eqn{\sum z \in \{1, p-1\}}.
+#'   The degree 1 hybrid will list those vectors and use them according
 #'   to their weights in the upcoming calculations. Depending on \eqn{p}, we can also go
 #'   a step further to a degree 2 hybrid by adding all \eqn{p(p-1)} vectors with
 #'   \eqn{\sum z \in \{2, p-2\}} to the process etc. The necessary predictions are
@@ -51,10 +41,9 @@
 #' 2. Step 2 (sampling part): The remaining weight is filled by sampling vectors z
 #'   according to Kernel SHAP weights normalized to the values not yet covered by Step 1.
 #'   Together with the results from Step 1 - correctly weighted - this now forms a
-#'   complete iteration as in CL21. The difference is that most mass is covered by exact
-#'   calculations. Afterwards, the algorithm iterates until convergence.
-#'   The output of Step 1 is reused in every iteration, leading to an extremely
-#'   efficient strategy.
+#'   complete iteration as in CL21. The difference is that a significant part of the mass
+#'   is covered by exact calculations. Afterwards, the algorithm iterates until
+#'   convergence. The output of Step 1 is reused in every iteration.
 #'
 #' If \eqn{p} is sufficiently small, all possible \eqn{2^p-2} on-off vectors \eqn{z} can be
 #' evaluated. In this case, no sampling is required and the algorithm returns exact
@@ -64,7 +53,7 @@
 #' should not be higher than 10 for exact calculations.
 #' For similar reasons, degree 2 hybrids should not use \eqn{p} larger than 40.
 #'
-#' @importFrom foreach %dopar%
+#' @importFrom doFuture %dofuture%
 #'
 #' @param object Fitted model object.
 #' @param X \eqn{(n \times p)} matrix or `data.frame` with rows to be explained.
@@ -82,12 +71,11 @@
 #'   Additional (named) arguments are passed via `...`.
 #'   The default, [stats::predict()], will work in most cases.
 #' @param feature_names Optional vector of column names in `X` used to calculate
-#'   SHAP values. By default, this equals `colnames(X)`. Not supported if `X`
-#'   is a matrix.
+#'   SHAP values. By default, this equals `colnames(X)`.
 #' @param bg_w Optional vector of case weights for each row of `bg_X`.
 #'   If `bg_X = NULL`, must be of same length as `X`. Set to `NULL` for no weights.
 #' @param bg_n If `bg_X = NULL`: Size of background data to be sampled from `X`.
-#' @param exact If `TRUE`, the algorithm will produce exact Kernel SHAP values
+#' @param exact If `TRUE`, the algorithm will produce exact SHAP values
 #'   with respect to the background data.
 #'   The default is `TRUE` for up to eight features, and `FALSE` otherwise.
 #' @param hybrid_degree Integer controlling the exactness of the hybrid strategy. For
@@ -97,12 +85,10 @@
 #'     worse than the hybrid strategy and should therefore only be used for
 #'     studying properties of the Kernel SHAP algorithm.
 #'   - `1`: Uses all \eqn{2p} on-off vectors \eqn{z} with \eqn{\sum z \in \{1, p-1\}}
-#'     for the exact part, which covers at least 75% of the mass of the Kernel weight
-#'     distribution. The remaining mass is covered by random sampling.
+#'     for the exact part. The remaining mass is covered by random sampling.
 #'   - `2`: Uses all \eqn{p(p+1)} on-off vectors \eqn{z} with
-#'     \eqn{\sum z \in \{1, 2, p-2, p-1\}}. This covers at least 92% of the mass of the
-#'     Kernel weight distribution. The remaining mass is covered by sampling.
-#'     Convergence usually happens in the minimal possible number of iterations of two.
+#'     \eqn{\sum z \in \{1, 2, p-2, p-1\}}. The remaining mass is covered by sampling.
+#'     Usually converges fast.
 #'   - `k>2`: Uses all on-off vectors with
 #'     \eqn{\sum z \in \{1, \dots, k, p-k, \dots, p-1\}}.
 #' @param m Even number of on-off vectors sampled during one iteration.
@@ -118,14 +104,18 @@
 #'   For `permshap()`, the default is 0.01, while for `kernelshap()` it is set to 0.005.
 #' @param max_iter If the stopping criterion (see `tol`) is not reached after
 #'   `max_iter` iterations, the algorithm stops. Ignored if `exact = TRUE`.
-#' @param parallel If `TRUE`, use parallel [foreach::foreach()] to loop over rows
-#'   to be explained. Must register backend beforehand, e.g., via 'doFuture' package,
-#'   see README for an example. Parallelization automatically disables the progress bar.
-#' @param parallel_args Named list of arguments passed to [foreach::foreach()].
-#'   Ideally, this is `NULL` (default). Only relevant if `parallel = TRUE`.
+#' @param parallel If `TRUE`, use [foreach::foreach()] and `%dofuture%` to loop over rows
+#'   to be explained. Must register backend beforehand, e.g., `plan(multisession)`,
+#'   see README for an example. Currently disables the progress bar.
+#' @param parallel_args Named list of arguments passed to
+#'   `foreach::foreach(.options.future = ...)`, ideally `NULL` (default).
+#'   Only relevant if `parallel = TRUE`.
 #'   Example on Windows: if `object` is a GAM fitted with package 'mgcv',
-#'   then one might need to set `parallel_args = list(.packages = "mgcv")`.
+#'   then one might need to set `parallel_args = list(packages = "mgcv")`.
+#'   Similarly, if the model has been fitted with `ranger()`, then it might be necessary
+#'   to pass `parallel_args = list(packages = "ranger")`.
 #' @param verbose Set to `FALSE` to suppress messages and the progress bar.
+#' @param seed Optional integer random seed. Note that it changes the global seed.
 #' @param survival Should cumulative hazards ("chf", default) or survival
 #'   probabilities ("prob") per time be predicted? Only in `ranger()` survival models.
 #' @param ... Additional arguments passed to `pred_fun(object, X, ...)`.
@@ -206,6 +196,7 @@ kernelshap.default <- function(
     parallel = FALSE,
     parallel_args = NULL,
     verbose = TRUE,
+    seed = NULL,
     ...) {
   p <- length(feature_names)
   basic_checks(X = X, feature_names = feature_names, pred_fun = pred_fun)
@@ -219,6 +210,10 @@ kernelshap.default <- function(
   bg_w <- prep_bg$bg_w
   bg_n <- nrow(bg_X)
   n <- nrow(X)
+
+  if (!is.null(seed)) {
+    set.seed(seed)
+  }
 
   # Calculate v1 and v0
   bg_preds <- align_pred(pred_fun(object, bg_X, ...))
@@ -236,12 +231,6 @@ kernelshap.default <- function(
   txt <- summarize_strategy(p, exact = exact, deg = hybrid_degree)
   if (verbose) {
     message(txt)
-  }
-
-  # Drop unnecessary columns in bg_X. If X is matrix, also column order is relevant
-  # In what follows, predictions will never be applied directly to bg_X anymore
-  if (!identical(colnames(bg_X), feature_names)) {
-    bg_X <- bg_X[, feature_names, drop = FALSE]
   }
 
   # Pre-calculations that are identical for each row to be explained
@@ -271,8 +260,9 @@ kernelshap.default <- function(
 
   # Apply Kernel SHAP to each row of X
   if (isTRUE(parallel)) {
-    parallel_args <- c(list(i = seq_len(n)), parallel_args)
-    res <- do.call(foreach::foreach, parallel_args) %dopar% kernelshap_one(
+    future_args <- c(list(seed = TRUE), parallel_args)
+    parallel_args <- c(list(i = seq_len(n)), list(.options.future = future_args))
+    res <- do.call(foreach::foreach, parallel_args) %dofuture% kernelshap_one(
       x = X[i, , drop = FALSE],
       v1 = v1[i, , drop = FALSE],
       object = object,
@@ -367,6 +357,7 @@ kernelshap.ranger <- function(
     parallel = FALSE,
     parallel_args = NULL,
     verbose = TRUE,
+    seed = NULL,
     survival = c("chf", "prob"),
     ...) {
   if (is.null(pred_fun)) {
@@ -389,6 +380,7 @@ kernelshap.ranger <- function(
     parallel = parallel,
     parallel_args = parallel_args,
     verbose = verbose,
+    seed = seed,
     ...
   )
 }